requirements.txt
Flask
torch
torchvision
Pillow
scikit-learn
numpy
